+++ src/main/java/duke/DukeEmptyArgumentException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DukeEmptyArgumentException extends DukeException{
     public DukeEmptyArgumentException(String msg) {
         super(msg);
