+++ src/main/java/duke/DukeFileNotFoundException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DukeFileNotFoundException extends DukeException{
     public DukeFileNotFoundException(String s) {
         super(s);
