+++ src/main/java/duke/DukeIOException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DukeIOException extends DukeException{
     public DukeIOException(String msg) {
         super(msg);
