+++ src/main/java/duke/DukeInvalidArgumentException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DukeInvalidArgumentException extends DukeException{
     public DukeInvalidArgumentException(String msg) {
         super(msg);
