+++ src/main/java/duke/DukeUnknownCommandException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DukeUnknownCommandException extends DukeException{
     public DukeUnknownCommandException(String msg) {
         super(msg);
