+++ src/main/java/duke/Task.java
@@ -1,3 +1,4 @@
+package duke;
 
 public class Task {
     protected String description;
