+++ src/main/java/duke/ToDos.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class ToDos extends Task{
     public ToDos(String description) {
         super(description);
