+++ src/test/java/duke/DeadlinesTest.java
@@ -1,3 +1,5 @@
+package duke;
+
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
